--- lib/fluent/plugin/opentelemetry/service_stub.rb
@@ -1,45 +1,0 @@
-# frozen_string_literal: true
-
-require "opentelemetry/proto/collector/logs/v1/logs_service_pb"
-require "opentelemetry/proto/collector/logs/v1/logs_service_services_pb"
-require "opentelemetry/proto/collector/metrics/v1/metrics_service_pb"
-require "opentelemetry/proto/collector/metrics/v1/metrics_service_services_pb"
-require "opentelemetry/proto/collector/trace/v1/trace_service_pb"
-require "opentelemetry/proto/collector/trace/v1/trace_service_services_pb"
-
-require "grpc"
-
-class Fluent::Plugin::Opentelemetry::ServiceStub
-  class Logs
-    def initialize(host, creds, **kw)
-      @stub = Opentelemetry::Proto::Collector::Logs::V1::LogsService::Stub.new(host, creds, **kw)
-    end
-
-    def export(json)
-      message = Opentelemetry::Proto::Collector::Logs::V1::ExportLogsServiceRequest.decode_json(json)
-      @stub.export(message)
-    end
-  end
-
-  class Metrics
-    def initialize(host, creds, **kw)
-      @stub = Opentelemetry::Proto::Collector::Metrics::V1::MetricsService::Stub.new(host, creds, **kw)
-    end
-
-    def export(json)
-      message = Opentelemetry::Proto::Collector::Metrics::V1::ExportMetricsServiceRequest.decode_json(json)
-      @stub.export(message)
-    end
-  end
-
-  class Traces
-    def initialize(host, creds, **kw)
-      @stub = Opentelemetry::Proto::Collector::Trace::V1::TraceService::Stub.new(host, creds, **kw)
-    end
-
-    def export(json)
-      message = Opentelemetry::Proto::Collector::Trace::V1::ExportTraceServiceRequest.decode_json(json)
-      @stub.export(message)
-    end
-  end
-end
