+++ lib/fluent/plugin/out_opentelemetry.rb
@@ -1,15 +1,17 @@
 # frozen_string_literal: true
 
 require "fluent/plugin/opentelemetry/constant"
-require "fluent/plugin/opentelemetry/request"
-require "fluent/plugin/opentelemetry/service_stub"
+require "fluent/plugin/opentelemetry/http_output_handler"
 require "fluent/plugin/output"
 
-require "excon"
-require "grpc"
 require "json"
-require "stringio"
-require "zlib"
+
+begin
+  require "grpc"
+
+  require "fluent/plugin/opentelemetry/grpc_output_handler"
+rescue LoadError
+end
 
 module Fluent::Plugin
   class OpentelemetryOutput < Output
@@ -57,12 +59,16 @@ module Fluent::Plugin
     def configure(conf)
       super
 
+      if @grpc_config && !defined?(GRPC)
+        raise Fluent::ConfigError, "To use gRPC feature, please install grpc gem such as 'fluent-gem install grpc'."
+      end
+
       unless [@http_config, @grpc_config].one?
         raise Fluent::ConfigError, "Please configure either <http> or <grpc> section."
       end
 
-      @http_handler = HttpHandler.new(@http_config, @transport_config, log) if @http_config
-      @grpc_handler = GrpcHandler.new(@grpc_config, @transport_config, log) if @grpc_config
+      @http_handler = Opentelemetry::HttpOutputHandler.new(@http_config, @transport_config, log) if @http_config
+      @grpc_handler = Opentelemetry::GrpcOutputHandler.new(@grpc_config, @transport_config, log) if @grpc_config
     end
 
     def multi_workers_ready?
@@ -80,127 +86,6 @@ module Fluent::Plugin
         @grpc_handler.export(record)
       else
         @http_handler.export(record)
-      end
-    end
-
-    class HttpHandler
-      def initialize(http_config, transport_config, logger)
-        @http_config = http_config
-        @transport_config = transport_config
-        @logger = logger
-
-        @tls_settings = {}
-        if @transport_config.protocol == :tls
-          @tls_settings[:client_cert] = @transport_config.cert_path
-          @tls_settings[:client_key] = @transport_config.private_key_path
-          @tls_settings[:client_key_pass] = @transport_config.private_key_passphrase
-          @tls_settings[:ssl_min_version] = Opentelemetry::TLS_VERSIONS_MAP[@transport_config.min_version]
-          @tls_settings[:ssl_max_version] = Opentelemetry::TLS_VERSIONS_MAP[@transport_config.max_version]
-        end
-
-        @timeout_settings = {
-          read_timeout: http_config.read_timeout,
-          write_timeout: http_config.write_timeout,
-          connect_timeout: http_config.connect_timeout
-        }
-      end
-
-      def export(record)
-        uri, connection = create_http_connection(record)
-        response = connection.post
-
-        if response.status != 200
-          if response.status == 400
-            # The client MUST NOT retry the request when it receives HTTP 400 Bad Request response.
-            raise Fluent::UnrecoverableError, "got unrecoverable error response from '#{uri}', response code is #{response.status}"
-          end
-
-          if @http_config.retryable_response_codes&.include?(response.status)
-            raise RetryableResponse, "got retryable error response from '#{uri}', response code is #{response.status}"
-          end
-          if @http_config.error_response_as_unrecoverable
-            raise Fluent::UnrecoverableError, "got unrecoverable error response from '#{uri}', response code is #{response.status}"
-          else
-            @logger.error "got error response from '#{uri}', response code is #{response.status}"
-          end
-        end
-      end
-
-      private
-
-      def http_logs_endpoint
-        "#{@http_config.endpoint}/v1/logs"
-      end
-
-      def http_metrics_endpoint
-        "#{@http_config.endpoint}/v1/metrics"
-      end
-
-      def http_traces_endpoint
-        "#{@http_config.endpoint}/v1/traces"
-      end
-
-      def create_http_connection(record)
-        msg = record["message"]
-
-        begin
-          case record["type"]
-          when Opentelemetry::RECORD_TYPE_LOGS
-            uri = http_logs_endpoint
-            body = Opentelemetry::Request::Logs.new(msg).body
-          when Opentelemetry::RECORD_TYPE_METRICS
-            uri = http_metrics_endpoint
-            body = Opentelemetry::Request::Metrics.new(msg).body
-          when Opentelemetry::RECORD_TYPE_TRACES
-            uri = http_traces_endpoint
-            body = Opentelemetry::Request::Traces.new(msg).body
-          end
-        rescue Google::Protobuf::ParseError => e
-          # The message format does not comply with the OpenTelemetry protocol.
-          raise ::Fluent::UnrecoverableError, e.message
-        end
-
-        headers = { Opentelemetry::CONTENT_TYPE => Opentelemetry::CONTENT_TYPE_PROTOBUF }
-        if @http_config.compress == :gzip
-          headers[Opentelemetry::CONTENT_ENCODING] = Opentelemetry::CONTENT_ENCODING_GZIP
-          gz = Zlib::GzipWriter.new(StringIO.new)
-          gz << body
-          body = gz.close.string
-        end
-
-        Excon.defaults[:ssl_verify_peer] = false if @transport_config.insecure
-        connection = Excon.new(uri, body: body, headers: headers, proxy: @http_config.proxy, persistent: true, **@tls_settings, **@timeout_settings)
-        [uri, connection]
-      end
-    end
-
-    class GrpcHandler
-      def initialize(grpc_config, transport_config, logger)
-        @grpc_config = grpc_config
-        @transport_config = transport_config
-        @logger = logger
-      end
-
-      def export(record)
-        msg = record["message"]
-
-        credential = :this_channel_is_insecure
-
-        case record["type"]
-        when Opentelemetry::RECORD_TYPE_LOGS
-          service = Opentelemetry::ServiceStub::Logs.new(@grpc_config.endpoint, credential)
-        when Opentelemetry::RECORD_TYPE_METRICS
-          service = Opentelemetry::ServiceStub::Metrics.new(@grpc_config.endpoint, credential)
-        when Opentelemetry::RECORD_TYPE_TRACES
-          service = Opentelemetry::ServiceStub::Traces.new(@grpc_config.endpoint, credential)
-        end
-
-        begin
-          service.export(msg)
-        rescue Google::Protobuf::ParseError => e
-          # The message format does not comply with the OpenTelemetry protocol.
-          raise ::Fluent::UnrecoverableError, e.message
-        end
       end
     end
   end
